fortify xml
